+++ react-components/src/components/Checkbox/model/public-api.ts
@@ -1,0 +1,1 @@
+export type { CheckboxGroupItem } from './checkbox-group-item';