+++ react-components/src/components/Table/index.tsx
@@ -2,44 +2,31 @@ import React, { useEffect, useMemo, useRef, useState } from 'react';
 import classNames from 'classnames';
 import {
   flexRender,
-  getCoreRowModel,
-  getSortedRowModel,
   Header,
   HeaderGroup,
   Row,
-  SortingState,
-  useReactTable,
+  Table as TanStackTable,
 } from '@tanstack/react-table';
 import { useVirtualizer } from '@tanstack/react-virtual';
-import objectHash, { sha1 } from 'object-hash';
-import dayjs from 'dayjs';
-import utc from 'dayjs/plugin/utc';
 
 import { TableProps } from './model';
 import { VerticalAlign } from './model/vertical-align';
 import TableRow from './components/row';
 import { Icon } from '../Icons';
 import { sortIconNames } from './sortIconNames';
-import { LocalStorageColumn, mergeSettings } from './storageUtils';
 import { Tooltip } from '../Tooltip';
-import { useColumnVisibility } from './useColumnVisibility';
-import {
-  debounce,
-  eventIsOnRow,
-  getCoordinates,
-  lockedClasses,
-} from './helpers';
-import { ICellEvent } from './model/i-cell-event';
-import { useTableColumns } from './useTableColumns';
-import { useLocalStorage } from '../../utils/useLocalStorage';
+import { Drawer } from '../Drawer';
+import { Button } from '../Button';
+import { Filter } from './components/filter';
+import { useTable } from './useTable';
+import { Stack } from '../Stack';
+import { Skeleton } from './components/skeleton';
+import { eventIsOnRow, getCoordinates, lockedClasses } from './helpers';
 import { LockedColumn } from './model/enum/locked-column.enum';
 import { useStickyStyles } from './useStickyStyles';
 
 import s from './style.module.scss';
 
-dayjs.extend(utc);
-
-const STORAGE_KEY = '_table_settings';
 const RESIZER_ATTRIBUTE_NAME = 'resizer';
 
 const verticalAlignValues: Record<VerticalAlign, string> = {
@@ -51,7 +38,7 @@ const verticalAlignValues: Record<VerticalAlign, string> = {
 export function Table<T>({
   dataSource,
   columns,
-  sticky = false,
+  sticky = true,
   skeleton = null,
   dictionary = null,
   cellParams = {
@@ -63,129 +50,37 @@ export function Table<T>({
   localStorageKey,
   className,
   valueChange,
-  dateFormat,
-  utcOffset,
-  roundToDecimalPlaces,
   verticalAlign = 'top',
+  headerComponent = null,
+  parent,
   ...props
 }: TableProps<T>): React.ReactElement {
-  // Ключ для localstorage
-  const storageKey = localStorageKey ? `${localStorageKey}${STORAGE_KEY}` : '';
-
   const currentEventRef = useRef<KeyboardEvent | MouseEvent | null>(null);
 
-  // Текущая сортировка столбцов
-  const [sorting, setSorting] = useState<SortingState>([]);
+  const [filterIsOpen, setFilterOpen] = useState(false);
 
-  // Выбранная строка
-  const [rowSelection, setRowSelection] = useState({});
+  let table: TanStackTable<T>;
 
-  // Текущая редактируемая ячейка
-  const [currentEditCell, setCurrentEditCell] = useState<ICellEvent | null>(
-    null,
-  );
+  if (props.table) {
+    table = props.table;
+  } else {
+    // Переделать conditional hook
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    table = useTable({
+      data: dataSource || [],
+      columns: columns || [],
+      dictionary,
+      storageKey: localStorageKey,
+    });
+  }
 
-  // Данные таблицы
-  const [data, setData] = useState<T[]>([...dataSource]);
-
-  useEffect(() => {
-    setData([...dataSource]);
-  }, [dataSource]);
-
-  // Настройки из localstorage
-  const [localStorageColumns, setLocalStorageColumns] = useLocalStorage<
-    LocalStorageColumn[]
-  >(
-    storageKey,
-    columns.map(({ name, width }) => ({
-      name,
-      width,
-    })),
-  );
-
-  const columnsHash = objectHash(columns.map((_) => _.name));
-
-  const mergedColumns = useMemo(
-    () => mergeSettings(columns, localStorageColumns),
-    [columnsHash],
-  );
-
-  // Скрытые колонки
-  const columnVisibility = useColumnVisibility(columns);
-
-  // Генерация колонок для react-table
-  const columnDefList = useTableColumns<T>({
-    columns: mergedColumns,
-    dictionary,
-    columnsHash,
-  });
-
-  const table = useReactTable({
-    data,
-    columns: columnDefList,
-    state: {
-      rowSelection,
-      sorting,
-      columnVisibility,
-    },
-    onSortingChange: setSorting,
-    enableRowSelection: true,
-    enableMultiRowSelection: false,
-    onRowSelectionChange: setRowSelection,
-    columnResizeMode: 'onChange',
-    getCoreRowModel: getCoreRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    meta: {
-      currentEditCell: currentEditCell,
-      startEditCell: (event: ICellEvent | null) => {
-        if (
-          currentEditCell?.column !== event?.column ||
-          currentEditCell?.row !== event?.row
-        ) {
-          setCurrentEditCell(event);
-        }
-      },
-      valueChanged: (value: T) => {
-        if (
-          currentEditCell?.row !== null &&
-          currentEditCell?.row !== undefined &&
-          currentEditCell?.column !== null
-        ) {
-          const row = table.getRow(currentEditCell.row.toString());
-          const column = table.getColumn(currentEditCell.column);
-          const foundRowIndex = data?.findIndex((_) => _ === row.original);
-
-          if (foundRowIndex !== -1) {
-            const updatedData = data.map((row, index) => {
-              if (foundRowIndex === index) {
-                return {
-                  ...value,
-                };
-              }
-
-              return row;
-            });
-
-            const isRowChanged = sha1(value!) !== sha1(data[foundRowIndex]!);
-
-            if (isRowChanged) {
-              setData(updatedData);
-
-              if (column && column.columnDef.meta) {
-                valueChange?.({
-                  row: updatedData[foundRowIndex],
-                  column: column.columnDef.meta.tableColumn,
-                });
-              }
-            }
-          }
-        }
-      },
-    },
+  table.options.meta?._onValueChanged?.((value: any) => {
+    valueChange?.(value);
   });
 
   const handleDblClick = (event: MouseEvent) => {
     const coordinates = getCoordinates(event);
+
     table.options?.meta?.startEditCell(coordinates);
   };
 
@@ -205,11 +100,13 @@ export function Table<T>({
     currentEventRef.current = event;
 
     if (event.key === 'Escape') {
-      setCurrentEditCell(null);
+      // setCurrentEditCell(null);
+      table.options.meta?.startEditCell(null);
     }
 
     if (event.key === 'Enter') {
-      setCurrentEditCell(null);
+      // setCurrentEditCell(null);
+      table.options.meta?.startEditCell(null);
     }
   };
 
@@ -264,13 +161,7 @@ export function Table<T>({
   // Сохранение в локальное хранилище
   const handleSaveColumns = (headers: Header<T, unknown>[]) => {
     if (!localStorageKey) return;
-
-    setLocalStorageColumns(
-      headers.map((item) => ({
-        name: item.id,
-        width: item.getSize(),
-      })),
-    );
+    table.options.meta?._saveColumnsToLocalStorage(headers);
   };
 
   // Сохранение после ресайза колонки
@@ -313,26 +204,6 @@ export function Table<T>({
     [tableWidth],
   );
 
-  const [horizontalScroll, setHorizontalScroll] = useState<{
-    left: number;
-    right: number;
-  }>({
-    left: 0,
-    right: 1,
-  });
-  const handleTableScroll = debounce(() => {
-    const maxScrollLeft =
-      (parentRef.current?.scrollWidth ?? 0) -
-      (parentRef.current?.offsetWidth ?? 0) +
-      12;
-    const scrollLeft = parentRef.current?.scrollLeft ?? 0;
-
-    setHorizontalScroll({
-      left: scrollLeft,
-      right: maxScrollLeft - scrollLeft,
-    });
-  }, 100);
-
   const cellStyles = {
     '--cell-vert-clamp': cellParams.verticalClamp,
     '--tbody-transform': `${virtualizer.getVirtualItems()[0]?.start}px`,
@@ -345,132 +216,164 @@ export function Table<T>({
     tableWidth,
   );
 
-  if (skeleton) return skeleton;
+  const state = table.getState();
+  const filterApplied = state.columnFilters.length > 0;
 
   return (
-    <div
-      className={s.root}
-      ref={parentRef}
-      onScroll={handleTableScroll}
-      style={{ height }}
-    >
-      <table
-        {...props}
-        className={classNames(s.table, className)}
-        style={{
-          ...cellStyles,
-          width: tableWidth,
-          height: virtualizer.getTotalSize(),
-        }}
+    <div className={s.root} style={{ height }}>
+      <Stack
+        size={12}
+        block
+        className={s.filters}
+        justifyContent="space-between"
       >
-        <thead className={classNames(sticky && s.sticky)}>
-          {table.getHeaderGroups().map((headerGroup) => (
-            <tr key={headerGroup.id}>
-              {headerGroup.headers.map((header) => {
-                const isSorted = header.column.getIsSorted();
-                const thContent = flexRender(
-                  header.column.columnDef.header,
-                  header.getContext(),
-                );
-                const thHint =
-                  columns.find(({ name }) => name === header.id)?.hint ?? '';
+        <div>{headerComponent}</div>
 
-                const columnLockedData = lockedColumnsVariables.get(
-                  header.column.id,
-                );
-                const columnLocked =
-                  columnLockedData?.lockedValue ?? LockedColumn.none;
+        <Stack divider>
+          <Button onClick={() => setFilterOpen(true)} view="ghost" square>
+            <Icon name={filterApplied ? 'filterApplied' : 'filter'} />
+          </Button>
+          <Button
+            disabled={!filterApplied}
+            onClick={() => table.resetColumnFilters()}
+            view="ghost"
+            square
+          >
+            <Icon name="filterClear" />
+          </Button>
+        </Stack>
+      </Stack>
+      <Drawer
+        open={filterIsOpen}
+        onClose={() => setFilterOpen(false)}
+        title="Фильтр"
+        parent={parent}
+        zIndex={2}
+        width={420}
+      >
+        <Filter
+          dict={dictionary}
+          table={table}
+          columns={table.getVisibleFlatColumns()}
+        />
+      </Drawer>
 
-                let stickyClasses = '';
-                if (columnLockedData?.isExtreme) {
-                  if (
-                    columnLocked === LockedColumn.left &&
-                    horizontalScroll.left !== 0
-                  ) {
-                    stickyClasses = s.lockedHeadLeftLast;
-                  } else if (
-                    columnLocked === LockedColumn.right &&
-                    horizontalScroll.right !== 0
-                  ) {
-                    stickyClasses = s.lockedHeadRightFirst;
-                  }
-                }
+      <div className={s.tableWrapper} ref={parentRef}>
+        <Skeleton skeleton={skeleton}>
+          <table
+            className={classNames(s.table, className)}
+            style={{
+              ...cellStyles,
+              width: tableWidth,
+              height: virtualizer.getTotalSize(),
+            }}
+          >
+            <thead className={classNames(sticky && s.sticky)}>
+              {table.getHeaderGroups().map((headerGroup) => (
+                <tr key={headerGroup.id}>
+                  {headerGroup.headers.map((header) => {
+                    const isSorted = header.column.getIsSorted();
+                    const thContent = flexRender(
+                      header.column.columnDef.header,
+                      header.getContext(),
+                    );
+                    const thHint =
+                      columns?.find(({ name }) => name === header.id)?.hint ??
+                      '';
+
+                    const columnLockedData = lockedColumnsVariables.get(
+                      header.column.id,
+                    );
+                    const columnLocked =
+                      columnLockedData?.lockedValue ?? LockedColumn.none;
+
+                    let stickyClasses = '';
+                    if (columnLockedData?.isExtreme) {
+                      if (columnLocked === LockedColumn.left) {
+                        stickyClasses = s.lockedHeadLeftLast;
+                      } else if (columnLocked === LockedColumn.right) {
+                        stickyClasses = s.lockedHeadRightFirst;
+                      }
+                    }
+
+                    return (
+                      <Tooltip
+                        key={header.id}
+                        title={thHint}
+                        target="hover"
+                        placement="bottom-start"
+                      >
+                        <th
+                          onClick={(event: React.MouseEvent<HTMLElement>) => {
+                            if (
+                              (event.target as HTMLElement).dataset.type ===
+                              RESIZER_ATTRIBUTE_NAME
+                            )
+                              return;
+
+                            const handler =
+                              header.column.getToggleSortingHandler();
+                            handler?.(event);
+                          }}
+                          className={classNames(
+                            lockedClasses[columnLocked]?.head,
+                            stickyClasses,
+                            header.column.getCanSort() && s.isSortable,
+                          )}
+                          style={columnLockedData?.variables}
+                        >
+                          {header.isPlaceholder ? null : (
+                            <div className={s.thContent}>
+                              <div>{thContent}</div>
+
+                              {isSorted ? (
+                                <Icon
+                                  name={sortIconNames.get(isSorted) ?? ''}
+                                  className={s.thContentIcon}
+                                />
+                              ) : null}
+                            </div>
+                          )}
+                          <div
+                            data-type={RESIZER_ATTRIBUTE_NAME}
+                            className={classNames(
+                              s.resizer,
+                              header.column.getIsResizing() &&
+                                s.resizerIsResizing,
+                            )}
+                            onMouseDown={(event) =>
+                              handleResizeMouseDown(event, header, headerGroup)
+                            }
+                          />
+                        </th>
+                      </Tooltip>
+                    );
+                  })}
+                </tr>
+              ))}
+            </thead>
+
+            <tbody>
+              {virtualizer.getVirtualItems().map((virtualRow) => {
+                const row = rows[virtualRow.index] as Row<T>;
 
                 return (
-                  <Tooltip
-                    key={header.id}
-                    title={thHint}
-                    target="hover"
-                    placement="bottom-start"
-                  >
-                    <th
-                      onClick={(event: React.MouseEvent<HTMLElement>) => {
-                        if (
-                          (event.target as HTMLElement).dataset.type ===
-                          RESIZER_ATTRIBUTE_NAME
-                        )
-                          return;
-
-                        const handler = header.column.getToggleSortingHandler();
-                        handler?.(event);
-                      }}
-                      className={classNames(
-                        lockedClasses[columnLocked]?.head,
-                        stickyClasses,
-                        header.column.getCanSort() && s.isSortable,
-                      )}
-                      style={columnLockedData?.variables}
-                    >
-                      {header.isPlaceholder ? null : (
-                        <div className={s.thContent}>
-                          <div>{thContent}</div>
-
-                          {isSorted ? (
-                            <Icon
-                              name={sortIconNames.get(isSorted) ?? ''}
-                              className={s.thContentIcon}
-                            />
-                          ) : null}
-                        </div>
-                      )}
-                      <div
-                        data-type={RESIZER_ATTRIBUTE_NAME}
-                        className={classNames(
-                          s.resizer,
-                          header.column.getIsResizing() && s.resizerIsResizing,
-                        )}
-                        onMouseDown={(event) =>
-                          handleResizeMouseDown(event, header, headerGroup)
-                        }
-                      />
-                    </th>
-                  </Tooltip>
+                  <TableRow
+                    key={virtualRow.key}
+                    virtualIndex={virtualRow.index}
+                    rowRef={virtualizer.measureElement}
+                    row={row}
+                    table={table}
+                    isSelectedRow={row.getIsSelected()}
+                    acrossLine={acrossLine}
+                    lockedColumnsVariables={lockedColumnsVariables}
+                  />
                 );
               })}
-            </tr>
-          ))}
-        </thead>
-
-        <tbody>
-          {virtualizer.getVirtualItems().map((virtualRow) => {
-            const row = rows[virtualRow.index] as Row<T>;
-
-            return (
-              <TableRow
-                key={virtualRow.key}
-                virtualIndex={virtualRow.index}
-                rowRef={virtualizer.measureElement}
-                row={row}
-                table={table}
-                isSelectedRow={row.getIsSelected()}
-                acrossLine={acrossLine}
-                lockedColumnsVariables={lockedColumnsVariables}
-                horizontalScroll={horizontalScroll}
-              />
-            );
-          })}
-        </tbody>
-      </table>
+            </tbody>
+          </table>
+        </Skeleton>
+      </div>
     </div>
   );
 }