+++ react-components/src/components/Table/model/public-api.ts
@@ -7,7 +7,6 @@ export * from './dictionary';
 export * from './cell-params';
 export * from './i-cell-component';
 export * from './i-property-grid';
-export * from './utc-offset';
 export * from './vertical-align';
 export * from './enum/locked-column.enum';
 export * from './cell-component';