+++ react-components/src/index.ts
@@ -17,8 +17,10 @@ export * from './components/Radio';
 export * from './components/Radio/model/public-api';
 
 export * from './components/Checkbox';
+export * from './components/Checkbox/model/public-api';
 
 export * from './components/Table';
+export { useTable } from './components/Table/useTable';
 export * from './components/Table/model/public-api';
 
 export * from './components/Dropdown';